moved game options to center view
