+++ MultiplayerExtensions/UI/Views/ClientLobbySetupPanel.bsml
@@ -1,0 +1,13 @@
+﻿<vertical>
+  <horizontal pad-top='40'>
+    <vertical min-width='45' horizontal-fit='MinSize'>
+      <toggle-setting id='CustomSongsToggle' value='CustomSongs' on-change='SetCustomSongs' text='Custom Songs' interactable='false' hover-hint='Whether the host has custom songs enabled'></toggle-setting>
+      <toggle-setting id='EnforceModsToggle' value='EnforceMods' on-change='SetEnforceMods' text='Enforce Mods' interactable='false' hover-hint='Whether players without MultiplayerExtensions will be kicked'></toggle-setting>
+    </vertical>
+    <vertical min-width='45' horizontal-fit='MinSize'>
+      <toggle-setting id="VerticalHUDToggle" value='VerticalHUD' on-change='SetVerticalHUD' text='Vertical HUD' hover-hint='Moves the HUD so that it is vertical'></toggle-setting>
+      <toggle-setting id="DefaultHUDToggle" value='DefaultHUD' on-change='SetDefaultHUD' text='Singleplayer HUD' hover-hint='Rearranges the HUD to look like it does in singleplayer'></toggle-setting>
+      <toggle-setting id="HologramToggle" value='Hologram' on-change='SetHologram' text='Hologram' hover-hint='Toggles the hologram showing the player in first place'></toggle-setting>
+    </vertical>
+  </horizontal>
+</vertical>
